// src/api/generated/model/ownershipTypeEnum.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */

/**
 * * `owned` - Owned
 * * `consigned` - Consigned
 * * `loaned` - Loaned
 */
export type OwnershipTypeEnum = typeof OwnershipTypeEnum[keyof typeof OwnershipTypeEnum];


export const OwnershipTypeEnum = {
  owned: 'owned',
  consigned: 'consigned',
  loaned: 'loaned',
} as const;